Version 3.1.2
Add dot visualization options to show the execution durations for all gathered inputs or as seperated values for each input.
Add timers for the state managers to show the time to acquire the state and the time to send the computed outputs to the following nodes
Fix type name when compiling with clang and msvc.